+++ prowler/providers/gcp/services/iam/iam_audit_logs_enabled/iam_audit_logs_enabled.py
@@ -14,13 +14,28 @@ class iam_audit_logs_enabled(Check):
                 project_id=project.id,
                 location=cloudresourcemanager_client.region,
             )
-            report.status = "PASS"
-            report.status_extended = f"Audit Logs are enabled for project {project.id}."
-            if not project.audit_logging:
-                report.status = "FAIL"
-                report.status_extended = (
-                    f"Audit Logs are not enabled for project {project.id}."
-                )
+            report.status = "FAIL"
+            report.status_extended = (
+                f"Default Audit Logs are not enabled for project {project.id}."
+            )
+            if project.audit_logging != []:
+                for policy in project.audit_logging:
+                    DEFAULT_AUDIT_LOGS = [
+                        "ADMIN_READ",
+                        "DATA_READ",
+                        "DATA_WRITE",
+                    ]
+                    if policy.get("service") == "allServices" and all(
+                        log
+                        in [
+                            log_config.get("logType")
+                            for log_config in policy.get("auditLogConfigs", [])
+                        ]
+                        for log in DEFAULT_AUDIT_LOGS
+                    ):
+                        report.status = "PASS"
+                        report.status_extended = f"Default Audit Logs are enabled for all services in project {project.id}."
+                        break
             findings.append(report)
 
         return findings